Transformando String em Num Decimal

+++ conceitos-javascript/script.js
@@ -1,5 +1,5 @@
 let n = '25.597';
 
-let res = parseInt(n); // Transformando String em Num Inteiro
+let res = parseFloat(n); // Transformando String em Num Decimal
 
 console.log(res);
